patch Ion cube in Player.Awake
