+++ ArchitectsLibrary/ArchitectsLibrary.csproj
@@ -217,6 +217,7 @@
     <Compile Include="Patches\BuilderPatches.cs" />
     <Compile Include="Patches\CraftingMenuPatches.cs" />
     <Compile Include="Patches\InspectOnFirstPickupPatches.cs" />
+    <Compile Include="Patches\PlayerPatches.cs" />
     <Compile Include="Patches\VehiclePatches.cs" />
     <Compile Include="Patches\WaterParkPatches.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
